add ratings migration

+++ src/main/java/com/sda/awesomemovies/api/ratings/RatingModel.java
@@ -5,12 +5,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
  class RatingModel {
     private Integer id;
-    private Integer rate;
+    private BigDecimal rate;
     private String author;
 }
